requirements.txt
pyyaml
pika
pytest